refactor: extract editableFlowJson module shared with global mode
